refactor: move destructure to dedicated module
